Move RegistrationsCard out of module
