+++ front/app/components/admin/GraphCards/OldRegistrationsByTimeCard/Chart.tsx
