+++ front/app/components/admin/GraphCards/OldRegistrationsByTimeCard/generateEmptyData.ts
