+++ front/app/components/admin/GraphCards/OldRegistrationsByTimeCard/index.tsx
