+++ front/app/components/admin/GraphCards/OldRegistrationsByTimeCard/useRegistrationsByTime/parse.ts
