+++ front/app/components/admin/GraphCards/OldRegistrationsByTimeCard/useRegistrationsByTime/query.ts
