+++ front/app/components/admin/GraphCards/OldRegistrationsByTimeCard/useRegistrationsByTime/translations.ts
