+++ front/app/components/admin/GraphCards/OldRegistrationsByTimeCard/useRegistrationsByTime/typings.ts
