+++ front/app/components/admin/GraphCards/RegistrationsCard/Chart.tsx
