+++ front/app/components/admin/GraphCards/RegistrationsCard/generateEmptyData.ts
