+++ front/app/components/admin/GraphCards/RegistrationsCard/index.tsx
