+++ front/app/components/admin/GraphCards/RegistrationsCard/messages.ts
