+++ front/app/components/admin/GraphCards/RegistrationsCard/renderTooltip.tsx
