+++ front/app/components/admin/GraphCards/RegistrationsCard/useRegistrations/index.ts
