+++ front/app/components/admin/GraphCards/RegistrationsCard/useRegistrations/messages.ts
