+++ front/app/components/admin/GraphCards/RegistrationsCard/useRegistrations/parse.ts
