+++ front/app/components/admin/GraphCards/RegistrationsCard/useRegistrations/translations.ts
