+++ front/app/components/admin/GraphCards/RegistrationsCard/useRegistrations/typings.ts
